Merge branch main into dev/exce_handling

+++ core/shared/utils/bh_assert.h
@@ -14,10 +14,10 @@ extern "C" {
 
 #if BH_DEBUG != 0
 void
-bh_assert_internal(int v, const char *file_name, int line_number,
+bh_assert_internal(int64 v, const char *file_name, int line_number,
                    const char *expr_string);
 #define bh_assert(expr) \
-    bh_assert_internal((int)(uintptr_t)(expr), __FILE__, __LINE__, #expr)
+    bh_assert_internal((int64)(uintptr_t)(expr), __FILE__, __LINE__, #expr)
 #else
 #define bh_assert(expr) (void)0
 #endif /* end of BH_DEBUG */
